Moved logout logic to new module, updated based on recommendations
