+++ Cargo.toml
@@ -51,7 +51,6 @@ serde_path_to_error = "0.1"
 serde_plain = "1.0"
 serde_with = "1.13"
 serde-value = "0.7"
-serde_urlencoded = "0.7.1"
 url = { version = "2.1", features = ["serde"] }
 subtle = "2.4"
