+++ src/discovery.rs
@@ -11,19 +11,12 @@ use serde::Serialize;
 use serde_with::{serde_as, skip_serializing_none, VecSkipError};
 use thiserror::Error;
 
-use crate::core::{
-    CoreAuthDisplay, CoreClaimName, CoreClaimType, CoreClientAuthMethod, CoreGrantType,
-    CoreJsonWebKey, CoreJsonWebKeyType, CoreJsonWebKeyUse, CoreJweContentEncryptionAlgorithm,
-    CoreJweKeyManagementAlgorithm, CoreJwsSigningAlgorithm, CoreResponseMode, CoreResponseType,
-    CoreSubjectIdentifierType,
-};
-
 use super::http_utils::{check_content_type, MIME_TYPE_JSON};
 use super::types::{
-    AuthDisplay, AuthenticationContextClass, ClaimName, ClaimType, ClientAuthMethod, EndSessionUrl,
-    GrantType, IssuerUrl, JsonWebKey, JsonWebKeySet, JsonWebKeySetUrl, JsonWebKeyType,
-    JsonWebKeyUse, JweContentEncryptionAlgorithm, JweKeyManagementAlgorithm, JwsSigningAlgorithm,
-    LanguageTag, OpPolicyUrl, OpTosUrl, RegistrationUrl, ResponseMode, ResponseType, ResponseTypes,
+    AuthDisplay, AuthenticationContextClass, ClaimName, ClaimType, ClientAuthMethod, GrantType,
+    IssuerUrl, JsonWebKey, JsonWebKeySet, JsonWebKeySetUrl, JsonWebKeyType, JsonWebKeyUse,
+    JweContentEncryptionAlgorithm, JweKeyManagementAlgorithm, JwsSigningAlgorithm, LanguageTag,
+    OpPolicyUrl, OpTosUrl, RegistrationUrl, ResponseMode, ResponseType, ResponseTypes,
     ServiceDocUrl, SubjectIdentifierType,
 };
 use super::{HttpRequest, HttpResponse, UserInfoUrl, CONFIG_URL_SUFFIX};
@@ -455,39 +448,6 @@ where
     #[error("Validation error: {0}")]
     Validation(String),
 }
-
-#[non_exhaustive]
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct LogoutProviderMetadata<A>
-where
-    A: AdditionalProviderMetadata,
-{
-    pub end_session_endpoint: Option<EndSessionUrl>,
-    #[serde(bound = "A: AdditionalProviderMetadata", flatten)]
-    pub additional_metadata: A,
-}
-impl<A> AdditionalProviderMetadata for LogoutProviderMetadata<A> where A: AdditionalProviderMetadata {}
-
-///
-/// Addition metadata for providers implementing OpenID Connect RP-Initiated Logout 1.0.
-///
-pub type ProviderMetadataWithLogout = ProviderMetadata<
-    LogoutProviderMetadata<EmptyAdditionalProviderMetadata>,
-    CoreAuthDisplay,
-    CoreClientAuthMethod,
-    CoreClaimName,
-    CoreClaimType,
-    CoreGrantType,
-    CoreJweContentEncryptionAlgorithm,
-    CoreJweKeyManagementAlgorithm,
-    CoreJwsSigningAlgorithm,
-    CoreJsonWebKeyType,
-    CoreJsonWebKeyUse,
-    CoreJsonWebKey,
-    CoreResponseMode,
-    CoreResponseType,
-    CoreSubjectIdentifierType,
->;
 
 #[cfg(test)]
 mod tests {
