+++ src/types.rs
@@ -548,7 +548,8 @@ new_type![
 
 new_url_type![
     ///
-    /// URL for the OpenID end session endpoint
+    /// URL for the [OpenID Connect RP-Initiated Logout 1.0](
+    /// https://openid.net/specs/openid-connect-rpinitiated-1_0.html) end session endpoint.
     ///
     EndSessionUrl
 ];
@@ -892,7 +893,7 @@ new_secret_type![
 
 new_secret_type![
     ///
-    /// Hint about the login identifier the End-User might use to log in.
+    /// Hint about the logout identifier the End-User might use to log out.
     ///
     /// The use of this parameter is left to the OpenID Connect Provider's discretion.
     ///
@@ -963,6 +964,15 @@ new_url_type![
     /// URL providing a client application's data usage policy.
     ///
     PolicyUrl
+];
+
+new_url_type![
+    ///
+    /// The post logout redirect URL, which should be passed to the end session endpoint
+    /// of providers implementing [OpenID Connect RP-Initiated Logout 1.0](
+    /// https://openid.net/specs/openid-connect-rpinitiated-1_0.html).
+    ///
+    PostLogoutRedirectUrl
 ];
 
 new_secret_type![
